Develop dcnb khdc

See merge request dev-teca-qlnv/qlnv-hang!1154

+++ src/main/java/com/tcdt/qlnvhang/table/dieuchuyennoibo/DcnbBienBanTinhKhoHdr.java
@@ -7,6 +7,7 @@ import lombok.*;
 
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -189,19 +190,19 @@ public class DcnbBienBanTinhKhoHdr extends BaseEntity implements Serializable, C
     private Long lanhDaoChiCucId;
 
     @Column(name = "TONG_SL_XUAT_THEO_QD")
-    private Double tongSlXuatTheoQd;
+    private BigDecimal tongSlXuatTheoQd;
 
     @Column(name = "TONG_SL_XUAT_THEO_TT")
-    private Double tongSlXuatTheoTt;
+    private BigDecimal tongSlXuatTheoTt;
 
     @Column(name = "SL_CON_LAI_THEO_SS")
-    private Double slConLaiTheoSs;
+    private BigDecimal slConLaiTheoSs;
 
     @Column(name = "SL_CON_LAI_THEO_TT")
-    private Double slConLaiTheoTt;
+    private BigDecimal slConLaiTheoTt;
 
     @Column(name = "CHENH_LECH_SL_CON_LAI")
-    private Double chenhLechSlConLai;
+    private BigDecimal chenhLechSlConLai;
 
     @Column(name = "TYPE")
     private String type;
